detect game-end; new game; view individual boards

+++ QuantumChess.App/Model/PieceProbability.cs
@@ -3,8 +3,8 @@
 	public class PieceProbability
 	{
 		public Piece Piece { get; set; }
-		public int Count { get; set; }
-		public int Potential { get; set; }
+		public long Count { get; set; }
+		public long Potential { get; set; }
 
 		public decimal Probability => Count / (decimal) Potential;
 	}
